mv(client/lib/utils): getInitials function from pages/Profile
